requirements.txt
Flask
SQLAlchemy
flask-sqlalchemy
openai
python-dotenv